2D scrolling now works fine using the Moose | Target indicators are correctly in place

+++ src/tools/Utils.java
@@ -52,7 +52,7 @@ public class Utils {
     public static double px2mm(double px) {
         String mName = cName + "px2mm";
 
-        return (int) ((px / PPI) * INCH_MM);
+        return (px / PPI) * INCH_MM;
     }
 
     /**
